+++ BiliAPI/BiliDynamic/DynamicModel/BiliDynamicDesc.cs
@@ -5,7 +5,7 @@
     /// </summary>
     public struct BiliDynamicDesc
     {
-        public int? uid { get; set; }
+        public long? uid { get; set; }
         public int? type { get; set; }
         public long? rid { get; set; }
         public int? acl { get; set; }
@@ -23,7 +23,7 @@
         public int? uid_type { get; set; }
         public int? stype { get; set; }
         public int? r_type { get; set; }
-        public int? inner_id { get; set; }
+        public long? inner_id { get; set; }
         public int? status { get; set; }
         public string? dynamic_id_str { get; set; }
         public string? pre_dy_id_str { get; set; }
