+++ BiliAPI/BiliUser/UserModel/BiliUserHonourInfo.cs
@@ -2,7 +2,7 @@
 {
     public struct BiliUserHonourInfo
     {
-        public int? mid { get; set; }
+        public long? mid { get; set; }
         public string? colour { get; set; }
         public string[] tags { get; set; }
     }
